// src/types/SubscriptionPlan.tsx
{/*  ===================================THIS FILE WAS AUTO GENERATED=================================== */}

export type SubscriptionPlan = {
    subscriptionPlanID: string;
    code: string;
    name: string;
    description: string;
    monthlyPrice: number;
    annualPrice: number;
    currency: string;
    maxProperties: number;
    maxRentalUnits: number;
    maxActiveListings: number;
    maxOrganizationMembers: number;
    features: string;
    providerProductID: string;
    providerMonthlyPriceID: string;
    providerAnnualPriceID: string;
    isActive: boolean;
    capturedDate: string;
    capturedBy: string;
    updatedDate: string;
    updatedBy: string;
};